requirements.txt
cassandra-driver
pymongo
mysql-connector-python
prettytable